+++ sdk/resourcemanager/azure-resourcemanager-resources/CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Added overload of `create(Context)` for `Creatable`.
 - Added overload of `apply(Context)` for `Appliable`.
+- Added overload of `beginCreate(Context)` for the definition flow of `Deployment`.
 - Added support for `resourceProviderName` and `featureName` in `Feature`.
 
 ## 2.6.0 (2021-06-18)